requirements.txt
Django==5.1.4
djangorestframework
requests
python-dotenv
django-cors-headers